requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
requests>=2.31,<3.0
pandas>=2.1,<3.0
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
docker-compose>=1.29,<2.0
drf-yasg>=1.21.7